✨ Separate out into context.rs and use properly from in the lib.rs
